Reform component structure
Compile with no errors. 
Linking with errors: 
undefined reference to `mysql_init@4' 
undefined reference to `mysql_real_connect@32' 
undefined reference to `mysql_error@4'
undefined reference to `mysql_errno@4'
undefined reference to `mysql_close@4'
undefined reference to `mysql_query@8'
undefined reference to `mysql_error@4'
undefined reference to `mysql_errno@4'
undefined reference to `mysql_close@4'
