+++ .vscode/c_cpp_properties.json
@@ -2,7 +2,10 @@
     "configurations": [{
         "name": "Win32",
         "includePath": [
-            "${workspaceFolder}/**"
+            "${workspaceFolder}/**",
+            "C:/Program Files/MariaDB/MariaDB Connector C 64-bit/lib",
+            "C:/Program Files/MariaDB/MariaDB Connector C 64-bit/include",
+            "C:/Program Files/MariaDB/MariaDB Connector C 64-bit"
         ],
         "defines": [
             "_DEBUG",
@@ -10,10 +13,10 @@
             "_UNICODE"
         ],
         "windowsSdkVersion": "10.0.17134.0",
-        "compilerPath": "C:/Program Files (x86)/Microsoft Visual Studio/2017/Community/VC/Tools/MSVC/14.15.26726/bin/Hostx64/x64/cl.exe",
+        "compilerPath": "C:/MinGW/bin/g++.exe",
         "cStandard": "c11",
         "cppStandard": "c++17",
-        "intelliSenseMode": "msvc-x64"
+        "intelliSenseMode": "gcc-x64"
     }],
     "version": 4
 }
